requirements.txt
Flask
google-cloud-vision
requests
python-dotenv
